backend/pyproject.toml
[project]
name = "bili-images-uploader"
version = "1.0.0"
description = "B站图片上传器"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pycryptodome>=3.20.0",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
]

[tool.uv]
dev-dependencies = []
